Fix bugs

+++ src/entities/Entity.java
@@ -21,7 +21,7 @@ public abstract class Entity {
         graphics.drawRect((int) hitbox.x, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
     }
 
-    protected void initHitbox(float x, float y, float width, float height) {    // int width, int height
+    protected void initHitbox(float x, float y, int width, int height) {    // int width, int height
         hitbox = new Rectangle2D.Float(x, y, width, height);
     }
